[FIX]Se mueve la función que hace scrapping.
Se mueve la función que hace el scrapping a la página del BCV de la
carpeta tootls a una función en el modelo res.company.
Adicionalmente se sube el timeout de 5 a 30 para pruebas.
Link:https://binaural.odoo.com/web#id=10125&cids=2&menu_id=293&action=393&active_id=58&model=helpdesk.ticket&view_type=form
